CheckScripObjStats
Checks the stats of the scriptable obj attached to the GameObject that ends up trigger the event which fires of an event with the UpdateStats function. This function increments the stats and gradually adds them to the StatText UI on screen.

IF YOU CHANGED YOUR HANDSDEMO SCENE STASH IT ADD IT BACK AND CHANGE THE NAME OF IT!