CoffeeFill Mechanic // Fixed Height Adjustment
Also added a couple lines of code in NewsSpawner to spawn a mug in case it gets destroyed by dropping it outside the map, to prevent the player getting stuck in the tutorial.

Height adjustment didn't work, since if you moved you would drop down to your original height. To fix this I applied the height transformation to the cameraoffset gameobject which holds the player camera.